Added colonies
First pass, mostly just reorganizing code. I still need to do some
serious cleanup around the way that creeps are spawned.
