+++ src/colonies/memory.d.ts
@@ -1,0 +1,13 @@
+// Copyright (c) 2018 Tim Perkins
+
+declare interface RoomMemory {
+  // Colony creep IDs
+  creepIds: string[];
+
+  // Colony source IDs
+  // sourceIds: string[];
+  // sourceIdPerCreepId: {[key: string]: string};
+
+  // TODO
+  // reservedRoomNames: string[];
+}
